feat: Add auth checks for process instance
