+++ db/rdbms/src/main/java/io/camunda/db/rdbms/read/service/VariableDbReader.java
@@ -45,7 +45,12 @@ public class VariableDbReader extends AbstractEntityReader<VariableEntity>
     final var dbSort = convertSort(query.sort(), VariableSearchColumn.VAR_KEY);
     final var dbQuery =
         VariableDbQuery.of(
-            b -> b.filter(query.filter()).sort(dbSort).page(convertPaging(dbSort, query.page())));
+            b ->
+                b.filter(query.filter())
+                    .authorizedResourceIds(resourceAccessChecks.getAuthorizedResourceIds())
+                    .authorizedTenantIds(resourceAccessChecks.getAuthorizedTenantIds())
+                    .sort(dbSort)
+                    .page(convertPaging(dbSort, query.page())));
     LOG.trace("[RDBMS DB] Search for variables with filter {}", query);
     final var totalHits = variableMapper.count(dbQuery);
     final var hits = variableMapper.search(dbQuery);
